==> 03.1_rothc_FB.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

Experiment: C-FB

#------------------------------------------------------------------------------#
# Script to run rothc with data from literature research and database: EJPSoil_BUFACAP_template.xlsx
#
# This script only treats experiment: Farako-Ba, Burkina Faso (SSA)
#
# created by T. Ruhnau (22.10.2024)
#------------------------------------------------------------------------------#

Read packages and input data -> data comes from 02_climate_data.R and 01_auto_database_FB.R
```{r Packages}
## load packages
library(SoilR)
library(readr)

# Read dataframes from both scripts before continue (first parts)
# > 02_climate_data_FB.R holds all important data on climate
# > 01_auto_database_FB.R holds all important data on soils
```
#-- Preperation ---------------------------------------------------------------#

The information of vegetation periods has to be taken directly from the literature or NDVI values, if not available in LTFEs

Split climate data up for historical spin-up (1970-1990) and prediction model (1991-2024)  
(The split-up assignment is random and follows no peer-reviewed basis)

According to Peña Cabriales et al. (2006) on pg. 59 it says: 
"one rainy season per year, starting in May-June and ending in October." 
> Filter is set accordingly.

Always run in the beginning
```{r Climate monthly means}
# Monthly means from climate_data_FB.R
mm1 <- read.csv2("output/farako_ba/climate/mm_1970_1990.csv", header = TRUE) #%>% select(-1)
mm2 <- read.csv2("output/farako_ba/climate/mm_1991_2024.csv", header = TRUE) #%>% select(-1)

# Filter Month and Experiment for Spinup section (1970 - 1990)
mm1_spinup <- mm1 %>%
  # filter(Month >= 5 & Month <= 10) %>% 
  filter(Experiment == "Farako-Ba") # Choose the experiment

# Filter Month and Experiment for Section1 Rain (1991 - 2024)
mm2_sec1 <- mm2 %>% # This will be used for Model Section
  filter(Month >= 5 & Month <= 10) %>% 
  filter(Experiment == "Farako-Ba") # Choose the experiment

# Filter Month and Experiment for Section2 Dry (1991 - 2024)
mm2_sec2 <- mm2 %>% # This will be used for Model Section
  filter(Month %in% c(11:12, 1:4)) %>% # to filter months through year
  filter(Experiment == "Farako-Ba") # Choose the experiment

rm(mm1, mm2)
```
```{r Read soil data}
# Specify the directory where your CSV files are located
csv_dir <- "output/farako_ba/" # change dir to different LTFEs

# Get a list of all CSV files in the directory
csv_files <- list.files(path = csv_dir, pattern = "*.csv", full.names = TRUE)

# Read all CSV files into a list of dataframes
df_list <- map(csv_files, ~read_csv2(.) %>% select(-1)) # -1 to get rid of first column

# Name the dataframes based on the file names (without extension)
names(df_list) <- tools::file_path_sans_ext(basename(csv_files))

# Unpack list to global environment
list2env(df_list, .GlobalEnv)

rm(df_list)
```

#-- Spin-up section -----------------------------------------------------------#

```{r Climate Spin-up}
## Climate data for Farako-ba, Burkina Faso for 1970-1990 (ERA5)

# Mean monthly temperature for 1970-1990
Temp_Spinup <- mm1_spinup %>% 
  filter(Variable == "Temp") %>% 
  select(Month, MeanValue) %>% 
  rename(Temp = MeanValue)

# Mean monthly precipitation for 1970-1990
Prec_Spinup <- mm1_spinup %>% 
  filter(Variable == "Prec") %>% 
  select(Month, MeanValue) %>% 
  rename(Prec = MeanValue)

# Mean monthly evaporation for 1970-1990
Evap_Spinup <- mm1_spinup %>% 
  filter(Variable == "Evapo") %>% 
  select(Month, MeanValue) %>% 
  rename(Evap = MeanValue)
```

```{r Soil Spin-up}
# Prepare soil data
soil.thick_Spinup=20 # Same as model
SOC_Spinup=10        # Change to fit initial SOC stock of model ---> 10 for C-FB
clay_Spinup=7        # Same as model
Cinputs_Spinup=1.774     # Change to historical land-use --> 1.744 for C-FB
# Former land use = Natural grassland
DPM.RPM = 0.67      # DPM/RPM --> DR = arable: 1.44, grasslands: 0.67, forest: 0.25 (Gottschalk et al. 2012)
FYM = 0              # Input factor for Farmyard manure by time

# Create vector of time steps to run spin-up
years_Spinup=seq(from=0, to=500, by=1/12)
```

```{r Prepare Spin-up}
# Calculate effects of climate on decomposition
fT_Spinup=fT.RothC(Temp_Spinup[,2]) # Temperature effects per month 
fW_Spinup=fW.RothC(P=(Prec_Spinup[,2]), E=(Evap_Spinup[,2]),
            S.Thick = soil.thick_Spinup, pClay = clay_Spinup,
            pE = 1.0, bare = FALSE)$b # Rainy season = vegetation covering the soil (bare = F)

# Replace negative values in fW with 0
fW_Spinup[fW_Spinup < 0] <- 0 # high Evapo values leads to error in model, still not sure if this is the right way to handle it

# >>Add comment<<
xi.frame_Spinup=data.frame(years_Spinup,rep(fT_Spinup*fW_Spinup,length.out=length(years_Spinup)))

# Estimating IOM using Falloon method
FallIOM=0.049*SOC_Spinup^(1.139) 
```

```{r Model Spin-up}
# Run the model Section1
Model_Spinup=RothCModel(t=years_Spinup,C0=c(DPM=0, RPM=0, BIO=0, HUM=0, IOM=FallIOM), 
                  In=Cinputs_Spinup, clay=clay_Spinup, xi=xi.frame_Spinup,
                  DR = DPM.RPM, FYM = FYM)

# Calculates SOC stocks for each pool
Ct_Spinup=getC(Model_Spinup) 

# Final pool size spin-up
poolSize_Spinup=as.numeric(tail(Ct_Spinup,1)) # Uses the last values to show final pool size
names(poolSize_Spinup)<-c("DPM", "RPM", "BIO", "HUM", "IOM") # Is that correct?

# Print pool size for spin-up section
poolSize_Spinup

# Sum the carbon stocks for total spin up
total_C_Spinup <- sum(poolSize_Spinup)
rows_C_Spinup <- rowSums(Ct_Spinup)

# Combine total carbon stocks into a data frame
total_C_df <- data.frame(
    Model = c("Spinup"),
    Total_C = c(sum(poolSize_Spinup))
)

# Print total sum of C after spin-up
print(total_C_df)
```

#-- Prediction section --------------------------------------------------------#

```{r Sec.1 Climate Model}
### Climate data for Farako-ba, Burkina Faso 
## Section1 Rainy Season (May-Oct) for 1991-2024 (ERA5)

# Mean monthly temperature for 1991-2024 (ERA5)
Temp_Model1 <- mm2_sec1 %>% 
  filter(Variable == "Temp") %>% 
  select(Month, MeanValue) %>% 
  rename(Temp = MeanValue)

# Mean monthly precipitation for 1991-2024 (ERA5)
Prec_Model1 <- mm2_sec1 %>% 
  filter(Variable == "Prec") %>% 
  select(Month, MeanValue) %>% 
  rename(Prec = MeanValue)

# Mean monthly evaporation for 1991-2024 (ERA5)
Evap_Model1 <- mm2_sec1 %>% 
  filter(Variable == "Evapo") %>% 
  select(Month, MeanValue) %>% 
  rename(Evap = MeanValue)
```

```{r Sec.2 Climate Model}
### Climate data for Farako-ba, Burkina Faso for 1991-2024 (ERA5)
## Section2 Dry Season (Nov-Apr) for 1991-2024 (ERA5)

# Mean monthly temperature for 1991-2024 (ERA5)
Temp_Model2 <- mm2_sec2 %>% 
  filter(Variable == "Temp") %>% 
  select(Month, MeanValue) %>% 
  rename(Temp = MeanValue)

# Mean monthly precipitation for 1991-2024 (ERA5)
Prec_Model2 <- mm2_sec2 %>% 
  filter(Variable == "Prec") %>% 
  select(Month, MeanValue) %>% 
  rename(Prec = MeanValue)

# Mean monthly evaporation for 1991-2024 (ERA5)
Evap_Model2 <- mm2_sec2 %>% 
  filter(Variable == "Evapo") %>% 
  select(Month, MeanValue) %>% 
  rename(Evap = MeanValue)
```

```{r Soil Prep C-FB}
### This chunk selects all parameters for the treatment C-FB (control site in Farako-Ba)
## For further information see literature in database

# Calculate soil.thick
fb_datasoil <- fb_datasoil %>% 
  mutate(soil.thick = Depth_to_cm - Depth_from_cm) # Indicates soil depth

# Get initial SOC stock 
fb_SOC <- fb_datasoil %>% 
  filter(tID == "C-FB") %>% # Change to respective experiment
  select(eID, tID, Sample_year, SOC_stock_Mg.ha) # Use first year as initial SOC stock, in this case 1993 = 16.06 Mg/ha

# Get clay content
fb_clay <- fb_soiltype %>% 
  filter(Bottom_depth_cm == 20) # Use value for clay model

# Get carbon residue input
fb_cinput <- fb_datacrop %>% 
  select(eID, tID, Residues_above_Mg.ha, DC_Comment) %>% 
  filter(tID == "C-FB") %>% # Change to respective experiment
  mutate(MeanResidues = mean(Residues_above_Mg.ha)) # Use mean for Cinputs in Model1
```

```{r Soil Model}
### Section1: Vegetation (Rainy Season) = TRUE (bare = FALSE)
## soil data
soil.thick_Model1 = 20    # Soil thickness (organic layer topsoil), in cm
SOC_Model1 = 10.22        # Spin-up should fit this initial SOC stock
clay_Model1 = 7           # Percent clay
Cinputs_Model1 = 2.510889 # 2.510889 for C-FB
Cinputs_Model2 = 0        # Annual C inputs to soil in Mg/ha/yr
DPM.RPM = 1.44            # DPM/RPM --> DR = arable: 1.44, grasslands: 0.67, forest: 0.25 (Gottschalk et al. 2012)
FYM = 0                   # FYM allocation: DPM: 49%, RPM: 49%, HUM: 2% (Weihermueller et al. 2013)

# Create vector of time steps to run model
years_Model1 = seq(1/12,10,by=1/12) # change sequence according to number of months in rain season
# years_Model2 = seq(1/6,20,by=1/6)
```

```{r Prepare Model}
#------------------------------- Section1 -------------------------------------#

# Calculate effects of climate on decomposition
fT_Model1=fT.RothC(Temp_Model1[,2]) # Temperature effects per month 
fW_Model1=fW.RothC(P=(Prec_Model1[,2]), E=(Evap_Model1[,2]), # Rain and evaporation effects
            S.Thick = soil.thick_Model1, pClay = clay_Model1,
            pE = 1.0, bare = FALSE)$b # Rainy season = vegetation covering the soil (bare = F)

# Replace negative values in fW with 0
fW_Model1[fW_Model1 < 0] <- 0 # high Evapo values leads to error in model

xi.frame_Model1=data.frame(years_Model1, rep(fT_Model1*fW_Model1, length.out=length(years_Model1)))

# #------------------------------- Section2 -------------------------------------#
# 
# # Calculate effects of climate on decomposition
# fT_Model2=fT.RothC(Temp_Model2[,2]) # Temperature effects 
# fW_Model2=fW.RothC(P=(Prec_Model2[,2]), E=(Evap_Model2[,2]), # Rain and evaporation effects
#             S.Thick = soil.thick_Model1, pClay = clay_Model1, # Same parameters as Section1
#             pE = 1.0, bare = TRUE)$b # Dry season = no vegetation covering the soil (bare = T)
# 
# # Replace negative values in fW with 0
# fW_Model2[fW_Model2 < 0] <- 0 # high Evapo values leads to error in model
# 
# # >>>ADD COMMENT<<
# xi.frame_Model2=data.frame(years_Model2, rep(fT_Model2*fW_Model2, length.out=length(years_Model2)))
# 
# #------------------------------------------------------------------------------#

# Estimating IOM using Falloon method
FallIOM=0.049*SOC_Model1^(1.139)
```

```{r Model Model}
#--------------------------- Section1: Run the model --------------------------#
# 
# Model_Model1=RothCModel(t=years_Model1,C0=poolSize_Spinup, # Using pool size from spin-up
#                   In=Cinputs_Model1, clay=clay_Model1, xi=xi.frame_Model1,
#                   DR = DPM.RPM, FYM = FYM)
# Ct_Model1=getC(Model_Model1) # Calculates SOC stocks for each pool

df_cinputs <- data.frame(
  month = years_Model1,
  values = c(0, 0, 0, 2.5, 2.5, 2.5, 2.5, 2.5, 2.5, 0, 0, 0)
)

df_fym <- data.frame(
  month = years_Model1,
  values = c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
)


test_model <- RothCModel(
  t=years_Model1,
  ks = c(k.DPM = 10, k.RPM = 0.3, k.BIO = 0.66, k.HUM = 0.02, k.IOM = 0),
  C0 = poolSize_Spinup,
  In = df_cinputs,
  FYM = df_fym,
  DR = DPM.RPM,
  clay = clay_Model1,
  xi = xi.frame_Model1,
  solver = deSolve.lsoda.wrapper,
  pass = FALSE
)

Ct_test=getC(test_model)
# #--------------------------- Section2: Run the model --------------------------#
# 
# Model_Model2=RothCModel(t=years_Model2,C0=poolSize_Spinup, # Using pool size from spin-up
#                   In=Cinputs_Model2, clay=clay_Model1, xi=xi.frame_Model2, # Only Cinput and xi.frame changes
#                   DR = DPM.RPM, FYM = FYM)
# Ct_Model2=getC(Model_Model2) # Calculates SOC stocks for each pool
# 
# #------------------------- Section1: Final pool size --------------------------#

poolSize_Model1=as.numeric(tail(Ct_Model1,1))
names(poolSize_Model1)<-c("DPM", "RPM", "BIO", "HUM", "IOM")
poolSize_Model1

poolSize_test=as.numeric(tail(Ct_test,1))
names(poolSize_test)<-c("DPM", "RPM", "BIO", "HUM", "IOM")
poolSize_test

# #------------------------- Section2: Final pool size --------------------------#
# 
# poolSize_Model2=as.numeric(tail(Ct_Model2,1))
# names(poolSize_Model2)<-c("DPM", "RPM", "BIO", "HUM", "IOM")
# poolSize_Model2
```

```{r Pool Model}
# #
# months <- rep(c(1:6), each = 10)
# years <- rep(c(1:10), each = 12)

# Sum of SOC stocks per row for both sections
## Consider seasonality: Model1 = rainy season and Model2 = dry season
rows_C_Model1 <- rowSums(Ct_Model1)
rows_C_test <- rowSums(Ct_test)

# # Combine both rows for months 1:4 & 5:10 & 11:12
# rows_C_total <- c(rows_C_Model2[1:80], rows_C_Model1, rows_C_Model2[81:120]) 
# rows_C_total2 <- c(rows_C_Model1, rows_C_Model2)

# # Merge results with months
# rows_months <- data.frame(
#   value = rows_C_Model1,
#   month = months
# )
# 
# # Merge results with months
# rows_years <- data.frame(
#   value = rows_C_Model1,
#   year = years
# )

# Sum of SOC stocks for both sections
## This is always at the end of predicition period = 20 years
total_C_Model1 <- sum(poolSize_Model1)
total_C_Model1 
total_C_test <- sum(poolSize_test)
total_C_test

# Create a data frame for the final pool sizes
final_pools_df <- data.frame(
    Pool = c("DPM", "RPM", "BIO", "HUM", "IOM"),
    Only_rainy = poolSize_Model1,
    All_season = poolSize_test,
    Spin_up = poolSize_Spinup
)

# Combine total carbon stocks into a data frame with spin-up
total_C_df <- data.frame(
    Model = c("Only_rainy", "All_season", "Spin_up"), #  
    Total_C = c(sum(poolSize_Model1), sum(poolSize_test), sum(poolSize_Spinup)), #
    Years = c(10, 10, 500) # 
    )

# Print results
print(total_C_df)

# Determine the maximum y-axis limit to fit both lines comfortably
max_C_Model1 <- max(total_C_Model1)
```

```{r Plot Model}
library(ggplot2)
library(gridExtra)

# Plot the final pool sizes
ggplot(final_pools_df, aes(x = Pool)) +
    geom_bar(aes(y = Only_rainy, fill = "Only_rainy"), stat = "identity", position = "dodge") +
    geom_bar(aes(y = All_season, fill = "All_season"), stat = "identity", position = "dodge") +
    labs(title = "Final Pool Sizes for Rainy_season and Dry_season",
         x = "Carbon Pools",
         y = "Carbon Stock (Mg/ha)",
         fill = "Model") +
    theme_minimal()

# Plot total carbon stocks
ggplot(total_C_df, aes(x = Model, y = Total_C, fill = Model)) +
    geom_bar(stat = "identity") +
    labs(title = "Total Carbon Stocks for Rainy_season and Dry_season",
         x = "Model",
         y = "Total Carbon Stock (Mg/ha)") +
    theme_minimal()

# Create a dataframe with rowsums and define time axis
df <- data.frame(
  time = c(1:6001, 6002:6061), #, 6002:6121, 6002:6121
  value = c(rows_C_Spinup, rows_C_Model1), #  rows_C_Model2, rows_C_total
  series = c(rep("Spin_up", 6001), rep("Rainy_season", 60)) # , rep("Dry_season", 120), rep("Total_C", 120)
)

# Create the plot with pre-defined time axis
ggplot(df, aes(x = time, y = value, color = series)) +
  geom_line() +
  scale_x_continuous(
    breaks = c(0, 6000, 6120),
    labels = c("-500", "0", "20"),
    limits = c(0, 6120)
  ) +
  labs(title = "Time Series Plot",
       x = "Time (years)",
       y = "Value") +
  theme_minimal(base_size = 20) +
  theme(legend.position = "bottom")

# Create a data frame with years, months, and values
df <- data.frame(
  year = rep(1:10, each = 6),
  month = rep(5:10, times = 10),
  values = rows_C_Model1
)

# Convert year and month to a proper date
df$date <- as.Date(paste(df$year, df$month, "01", sep = "-"))

# Create a data frame for the Farako-ba points
farako_ba <- data.frame(
  eID = rep("Farako-Ba", 3),
  tID = rep("C-FB", 3),
  date = as.Date(c("0001-05-01", "0003-05-01", "0005-05-01")),
  values = c(10.2, 10.8, 9.34)
)

# Create the ggplot
ggplot() +
  geom_line(data = df, aes(x = date, y = values), color = "blue", size = 1) +
  geom_point(data = farako_ba, aes(x = date, y = values), color = "green", size = 5) +
  geom_text(data = farako_ba, aes(x = date, y = values, label = paste(eID, values)), 
            vjust = -1.5, hjust = 0.5, color = "black", fontface = "bold") +
  scale_x_date(date_breaks = "1 years", date_labels = "%Y", 
               limits = as.Date(c("0001-05-01", "0010-10-01"))) +
  scale_y_continuous(breaks = seq(0, 60, by = 0.5)) +
  labs(title = "C-FB observation vs prediction",
       x = "Year",
       y = "SOC stock (Mg/ha)") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        legend.position = "bottom")
```

#-- Dump it --------------------------------------------------------------------

```{r Dumpster}
#!# values from ERA5 tiff are all negative, calculation error?
## seems like evaporation = negative values and condensation = positive values according to ERA5 information
## all values are in m > *-1000 to convert to positive values in mm

# Original values from ERA5
# Evp=data.frame(Month=1:12, Evp=c(-0.3, -0.3, -0.3, -0.3, -0.26, -0.24, -0.19, -0.15, -0.15, -0.22, -0.26, -0.32))

# ## plot results
# matplot(years_Spinup, Ct_Spinup, type="l", lty=1, col=1:5,
#         xlab="Time (years)", ylab="C stocks (Mg/ha)")
# legend("topleft", c("DPM", "RPM", "BIO", "HUM", "IOM"),
#        lty=1, col=1:5, bty="n")

# 
# # Plot individual pools and total SOC for the spin-up period
# matplot(years_, Ct_Spinup, type = "l", lty = 1, col = 1:5,
#         xlab = "Time (years)", ylab = "C stocks (Mg/ha)",
#         main = "SOC Pools and Total (Spin-Up)", ylim = c(0, max_C_Model1 * 1.1))
# 
# lines(years_Spinup, total_C_spinup, col = "black", lty = 2)
# legend("topright", c("Total SOC", "DPM", "RPM", "BIO", "HUM", "IOM"),
#        lty = c(1, 1, 1, 1, 1, 1), col = c(1:5, "black"), bty = "n")

# # Plot 1: Line plot
# plot1 <- ggplot(rows_years, aes(x = year, y = value, group = 1)) +
#   geom_line() +
#   geom_point() +
#   theme_minimal() +
#   labs(title = "Plot of SOC stock by years",
#        x = "Years",
#        y = "SOC Stock (Mg/ha)")
# 
# # Plot 2: Box plot
# plot2 <- ggplot(result_df, aes(x = month, y = value)) +
#   geom_boxplot() +
#   theme_minimal() +
#   labs(title = "Box Plot of Values by Month",
#        x = "Month",
#        y = "Value")
# 
# # Combine plots
# combined_plot <- grid.arrange(plot1, plot2, ncol = 1)

```
