==> 04_visualize.Rmd <==
---
title: "04_visualize"
author: "Tobias Ruhnau"
date: "2024-11-06"
output: html_document
---
#--- Preperation

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r libraries}
library(ggplot2)
library(gridExtra)
library(tidyr)
```

#--- 01_import.Rmd

```{r plot import, echo=FALSE}
## Load dataframes from 01_import.Rmd

# Calculate the overall min and max of SOC_stock
y_min <- min(ndp_datasoil$SOC_stock_Mg.ha, na.rm = TRUE)
y_max <- max(ndp_datasoil$SOC_stock_Mg.ha, na.rm = TRUE)

# Calculate a reasonable step size (e.g., 1/10th of the range)
y_step <- (y_max - y_min) / 10

# Create a list of plots, one for each treatment
plot_list <- ndp_datasoil %>%
  split(.$tID) %>% # split plots for each treatment
  map(function(data) {
    ggplot(data, aes(x = Sample_year, y = SOC_stock_Mg.ha, group = 1)) +
      geom_line() +
      geom_point() +
      theme_minimal(base_line_size = 20) +
      labs(title = paste("Treatment:", unique(data$tID)),
           x = "Year",
           y = "SOC stock (Mg/ha)") +
      theme(axis.text.x = element_text(angle = 45, hjust = 1),
            plot.title = element_text(size = 10),
            axis.title = element_text(size = 9),
            axis.text = element_text(size = 8)) +
      scale_y_continuous(
        limits = c(y_min, y_max),
                         breaks = seq(y_min, y_max, by = y_step)) +  # Set common y-axis limits
      scale_x_continuous(
        breaks = function(x) seq(ceiling(min(x)), floor(max(x)), by = 1), # Change according to years (1 = every year)
        labels = function(x) format(x, nsmall = 0)
      )
  })


# please note: C2-FB shows the natural grassland and hence, has the highest SOC stock (no conversion)
# please note: Nioro-du-rip has always the same SOC stock measured -> should be checked!
combined_plot <- do.call(grid.arrange, c(plot_list, ncol = length(plot_list)))
 
# ## If necessary, this helps!
# # Arrange the plots in a 2x5 grid (5 plots in first row, 4 in second)
# combined_plot <- grid.arrange(
#   grobs = plot_list,
#   layout_matrix = rbind(1:5, c(6:9, NA)),
#   widths = rep(1, 5),
#   heights = c(1, 1)
# )

# Display the combined plot
plot(combined_plot)

# # Save the plot
# ggsave("C:/Users/t.ruhnau/OneDrive - Unique/Dokumente/plots/nioro_du_rip/SOC_treatments_NDP.png",
#        combined_plot, width = 20, height = 10, units = "in") # Create folder manually
```

#--- 03.1.1.rothc_C-FB.Rmd

```{r C-FB pool sizes, echo=FALSE}
# Plot the final pool sizes
ggplot(FB_C_final_pools, aes(x = Pool)) +
    geom_bar(aes(y = Model, fill = "Model"), stat = "identity", position = "dodge") +
    labs(title = "C-FB Final Pool Sizes",
         x = "Carbon Pools",
         y = "Carbon Stock (Mg/ha)",
         fill = "Model") +
    theme_minimal()
```

```{r C-FB total C, echo=FALSE}
# Plot total carbon stocks
ggplot(FB_C_total_C, aes(x = Model, y = Total_C, fill = Model)) +
    geom_bar(stat = "identity") +
    labs(title = "C-FB Total Carbon Stocks",
         x = "Model",
         y = "Total Carbon Stock (Mg/ha)") +
    theme_minimal()
```

```{r C-FB spin-up & pred, echo=FALSE}
# Create a dataframe with rowsums and define time axis
df <- data.frame(
  time = c(1:6001, 6002:6241),
  value = c(FB_C_rows_C_Spinup, FB_C_rows_C_Model),
  series = c(rep("Spin_up", 6001), rep("Prediction", 240))
)

# Create the plot with pre-defined time axis
ggplot(df, aes(x = time, y = value, color = series)) +
  geom_line() +
  scale_x_continuous(
    breaks = c(0, 6000, 6241),
    labels = c("-500", "0", "20"),
    limits = c(0, 6241)
  ) +
  labs(title = "C-FB Spin-up & Prediction",
       x = "Years",
       y = "Value") +
  theme_minimal(base_size = 20) +
  theme(legend.position = "bottom")

# # Save the plot
# ggsave("pools_model_FYM-FB", device="png", "plots/farako_ba/treatments/", width = 10, height = 6)
```

```{r C-FB obs vs pred, echo=FALSE}
# Create a data frame with years, months, and values
df1 <- data.frame(
  year = rep(1:10, each = 6), # edit
  month = rep(5:10, times = 10), # edit
  values = FB_C_rows_C_Model
)

# Convert year and month to a proper date
df1$date <- as.Date(paste(df$year, df$month, "01", sep = "-"))

# Create a data frame for the Farako-ba points
C_FB <- data.frame(
  eID = rep("Farako-Ba", 3),
  tID = rep("C-FB", 3),
  date = as.Date(c("0001-05-01", "0003-05-01", "0005-05-01")),
  values = c(10.2, 10.8, 9.34) # Exact values from literature found in fb_SOC
)

# Create the ggplot
ggplot() +
  geom_line(data = df1, aes(x = date, y = values), color = "blue", size = 1) +
  geom_point(data = C_FB, aes(x = date, y = values), color = "green", size = 5) +
  geom_text(data = C_FB, aes(x = date, y = values, label = paste(eID, values)), 
            vjust = -1.5, hjust = 0.5, color = "black", fontface = "bold") +
  scale_x_date(date_breaks = "1 years", date_labels = "%Y", 
               limits = as.Date(c("0001-05-01", "0010-10-01"))) + # edit
  scale_y_continuous(breaks = seq(0, 60, by = 0.5)) +
  labs(title = "C-FB Observation vs. Prediction",
       x = "Year",
       y = "SOC stock (Mg/ha)") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        legend.position = "bottom")

# # Save the plot
# ggsave("pools_model_FYM-FB", device="png", "plots/farako_ba/treatments/", width = 10, height = 6)
```

#--- 03.1.2.rothc_FYM-FB.Rmd

```{r FYM-FB final pools, echo=FALSE}
# Plot the final pool sizes
ggplot(FB_FYM_final_pools, aes(x = Pool)) +
    geom_bar(aes(y = Model, fill = "Model"), stat = "identity", position = "dodge") +
    labs(title = "FYM-FB Final Pool Sizes",
         x = "Carbon Pools",
         y = "Carbon Stock (Mg/ha)",
         fill = "Model") +
    theme_minimal()

# # Save the plot
# ggsave("pools_model_FYM-FB", device="png", "plots/farako_ba/treatments/", width = 10, height = 6)
```

```{r FYM-FB total C, echo=FALSE}
# Plot total carbon stocks
ggplot(FB_FYM_total_C, aes(x = Model, y = Total_C, fill = Model)) +
    geom_bar(stat = "identity") +
    labs(title = "FYM-FB Total Carbon Stocks",
         x = "Model",
         y = "Total Carbon Stock (Mg/ha)") +
    theme_minimal()

# # Save the plot
# ggsave("pools_model_FYM-FB", device="png", "plots/farako_ba/treatments/", width = 10, height = 6)
```

```{r FYM-FB spin-up & pred, echo=FALSE}
# Create a dataframe with rowsums and define time axis
df <- data.frame(
  time = c(1:6001, 6002:6241),
  value = c(FB_FYM_rows_C_Spinup, FB_FYM_rows_C_Model),
  series = c(rep("Spin_up", 6001), rep("Prediction", 240))
)

# Create the plot with pre-defined time axis
ggplot(df, aes(x = time, y = value, color = series)) +
  geom_line() +
  scale_x_continuous(
    breaks = c(0, 6000, 6241),
    labels = c("-500", "0", "20"),
    limits = c(0, 6241)
  ) +
  labs(title = "FYM-FB Spin-up & Prediction",
       x = "Years",
       y = "Value") +
  theme_minimal(base_size = 20) +
  theme(legend.position = "bottom")

# # Save the plot
# ggsave("pools_model_FYM-FB", device="png", "plots/farako_ba/treatments/", width = 10, height = 6)
```

```{r FYM-FB obs vs pred}
# Create a data frame with years, months, and values
df1 <- data.frame(
  year = rep(1:20, each = 12), # Year = July to June
  month = rep(1:12, times = 20), # Month 1 = July, Month 12 = June
  values = FB_FYM_rows_C_Model
)

# Convert year and month to a proper date
df1$date <- as.Date(paste(df1$year, df1$month, "01", sep = "-"))

# Create a data frame for the Farako-ba points
FYM_FB <- data.frame(
  eID = rep("Farako-Ba", 3),
  tID = rep("PK+ME-FB", 3),
  date = as.Date(c("0001-05-01", "0004-05-01", "0007-05-01")), # Years between each measurement
  values = c(12, 13.1, 12.8) # Exact values from literature found in fb_SOC
)

# Create the ggplot
ggplot() +
  geom_line(data = df1, aes(x = date, y = values), color = "blue", size = 1) +
  geom_point(data = FYM_FB, aes(x = date, y = values), color = "green", size = 5) +
  geom_text(data = FYM_FB, aes(x = date, y = values, label = paste(eID, values)), 
            vjust = -1.5, hjust = 0.5, color = "black", fontface = "bold") +
  scale_x_date(date_breaks = "1 years", date_labels = "%Y", 
               limits = as.Date(c("0001-01-01", "0020-06-01"))) +
  scale_y_continuous(breaks = seq(0, 60, by = 0.5)) +
  labs(title = "FYME-FB Observation vs. Prediction",
       x = "Year",
       y = "SOC stock (Mg/ha)") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        legend.position = "bottom")

# # Save the plot
# ggsave("pools_model_FYM-FB", device="png", "plots/farako_ba/treatments/", width = 10, height = 6)
```
